% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_installed.R
\name{gs_check_installed}
\alias{gs_check_installed}
\title{Check whether a package is installed.}
\usage{
gs_check_installed(pkg, strict = TRUE, prompt = base::interactive(), ...)
}
\arguments{
\item{pkg}{A character vector naming the package(s), whose installation
needs to be checked in any of the libraries. Note that the character vector
should be length 1.}

\item{strict}{If \code{strict} is \code{TRUE} and the package is not installed, the
function with throw an error. Otherwise, a \code{logical} value is returned.}

\item{prompt}{If \code{TRUE}, will prompt the user to install needed package.}

\item{...}{Currently ignored}
}
\value{
If \code{strict = TRUE}, and \code{package} is not yet installed, the
function stops and throws an error. Else, a logical vector is
returned, indicating whether the package is installed.
}
\description{
Check whether a package is installed.
}
